Extracted premiums and favorites controllers
