+++ src/modules/favorites/favorites.const.ts
@@ -1,0 +1,4 @@
+export enum FavoritesAction {
+  Add = 1,
+  Remove = 0,
+}
