+++ src/modules/offer/offer.const.ts
@@ -1,6 +1,2 @@
 export const DEFAULT_OFFER_COUNT = 60;
 export const PREMIUM_OFFER_COUNT = 3;
-export enum FavoritesAction {
-  Add = 1,
-  Remove = 0,
-}
