+++ src/modules/premiums/premiums.controller.ts
@@ -2,9 +2,9 @@ import { Controller, LoggerInterface } from '../../common/index.js';
 import { inject, injectable } from 'inversify';
 import { Component, HttpMethod } from '../../types/index.js';
 import { Request, Response } from 'express';
-import { OfferServiceInterface } from './offer-service.interface.js';
+import { OfferServiceInterface } from '../offer/offer-service.interface.js';
 import { fillDTO } from '../../utils/index.js';
-import OfferShortResponse from './response/offer-short.response.js';
+import OfferShortResponse from '../offer/response/offer-short.response.js';
 
 @injectable()
 export default class PremiumsController extends Controller {
